Changed folder structure and updated web pages
Changed folder structure to be better understood by iGEM-team.

Updated Home.html and Description.html with respect to current page on iGEM.

COMMENT: Reconsider edit on project description. The text in this latest version is quite long, making it daunting to read, as well as exceeding judges criteria. If more details are desired these should rather be put under "Se more" tabs or in the separate page "Background".